requirements.txt
alembic==1.8.1
astroid==2.12.12
asttokens==2.0.8
backcall==0.2.0
click==8.1.3
colorama==0.4.6
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
executing==1.1.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
gunicorn==20.1.0
ipykernel==6.16.2
ipython==8.5.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jotform==1.2
jupyter_client==7.4.4
jupyter_core==4.11.2
lazy-object-proxy==1.7.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.3
pure-eval==0.2.2
Pygments==2.13.0
pylint==2.15.5
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pyzmq==24.0.1
six==1.16.0
SQLAlchemy==1.4.42
stack-data==0.5.1
tomli==2.0.1
tomlkit==0.11.5
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
